requirements.txt
discord
python-dotenv
git+https://github.com/yuce/pyswip@master#egg=pyswip